Version 3.0/requirements.txt
numpy
tensorflow
matplotlib
seaborn
scikit-learn
flask
pillow
